Update patch set 9
Patch Set 9:

Hey guys! Please pardon me for posting this here as it would be too much if posted on G+. Anyway, the battery alignment is still off with the latest release (BETA 0.20).

I've captured a screenshot here and a few ms paint mastery.
https://drive.google.com/open?id=0ByF8eblo4zMGUWtlR1Z5ZGpweGs

As you can see, there's two red line below and above the statusbar icons. There's nothing wrong with the above line as the 3 icons are aligned with "1 pixel" each I put in there (as an indication that they touched the icons). However, below seems to be a different case, the wifi icon for example takes around "3 pixel" while the signal icon takes "2 pixel" and lastly, the battery takes "1 pixel" to touch the line. The result is this makes the battery look bigger. I appreciate you reading up to this point.

Btw, I used bacon with a clean flash and if you need me for testing purposes, I'd be very happy to help. Thanks!

Patch-set: 9